% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{addFitSummary}
\alias{addFitSummary}
\title{Add model summary to an object of class mySummary}
\usage{
addFitSummary(df, fit, statsname = "", saveid = FALSE)
}
\arguments{
\item{df}{AN object of class mySummary}

\item{fit}{An object of class glm or lm}

\item{statsname}{character Name of statistics}

\item{saveid}{logical whether or not save id}
}
\description{
Add model summary to an object of class mySummary
}
\examples{
require(lme4)
data(colon_s,package="finalfit")
fit=glm(mort_5yr~age.factor+sex.factor+obstruct.factor+perfor.factor,data=colon_s,family="binomial")
df=autoReg(fit,uni=TRUE,multi=TRUE,threshold=1,saveid=TRUE)
fit2=lme4::glmer(mort_5yr~age.factor+obstruct.factor+(1|hospital),data=colon_s,family="binomial")
df \%>\% addFitSummary(fit2,statsname="OR (multilevel)") \%>\% myft()
}
