% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabelData.R
\name{addLabelData}
\alias{addLabelData}
\title{Add labels to data}
\usage{
addLabelData(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame
}
\description{
Add labels to data
}
\examples{
addLabelData(data.frame(ph.ecog=0:3,sex=c(1,2,2,2),age=c(20,30,40,70)))
}
