% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPlot.R
\name{adjustedPlot2}
\alias{adjustedPlot2}
\title{Draw a survfitted plot}
\usage{
adjustedPlot2(fit, se = FALSE, mark.time = FALSE)
}
\arguments{
\item{fit}{An object of class coxph or survfit}

\item{se}{logical Whether or not show se}

\item{mark.time}{logical Whether or not mark time}
}
\value{
a ggplot
}
\description{
Draw a survfitted plot
}
\examples{
library(survival)
fit=coxph(Surv(time,status)~rx+logWBC,data=anderson)
adjustedPlot2(fit)
}
