% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson.R
\docType{data}
\name{anderson}
\alias{anderson}
\title{Remission survival times of 42 leukemia patients}
\format{
A data.frame with 42 rows and 5 variables
\describe{
\item{time}{survival time in weeks}
\item{status}{censoring status 1=failure 0=censored}
\item{sex}{sex 1=Female 0=Male}
\item{logWBC}{log white blood cell count}
\item{rx}{treatment status 1=control 0 =treatment}
}
}
\source{
David G. Kleinbaum and Mitchel Klein. Survival Analysis. A Self-Learning Text(3rd ed)(Springer,2012) ISBN: 978-1441966452
}
\usage{
anderson
}
\description{
A dataset containing  survival time of 42 leukemia patients
}
\keyword{datasets}
