% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson.R
\docType{data}
\name{anderson2}
\alias{anderson2}
\title{Remission survival times of 31 leukemia patients}
\format{
A data.frame with 31 rows and 6 variables
\describe{
\item{time}{survival time in weeks}
\item{status}{censoring status 1=failure 0=censored}
\item{sex}{sex 0=Female 1=Male}
\item{logWBC}{log white blood cell count}
\item{rx}{treatment status 1=control 0 =treatment}
\item{WBCCAT}{WBC count group 1=medium 2=high}
}
}
\source{
David G. Kleinbaum and Mitchel Klein. Survival Analysis. A Self-Learning Text(3rd ed)(Springer,2012) ISBN: 978-1441966452
}
\usage{
anderson2
}
\description{
This data is subdata of anderson with medium(2.3 < logWBC <= 2.96) and high WBC count(logWBC > 2.96)
}
\details{
A dataset containing  survival time of 31 leukemia patients
}
\keyword{datasets}
