% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_printable.R
\name{as_printable}
\alias{as_printable}
\title{Convert data.frame to printable form}
\usage{
as_printable(
  data,
  align.first = "left",
  align.chr = "right",
  align.num = "right"
)
}
\arguments{
\item{data}{A data.frame}

\item{align.first}{character Alignment of first variable}

\item{align.chr}{character Alignment of character variable}

\item{align.num}{character Alignment of numeric variable}
}
\value{
A data.frame
}
\description{
Calculate character length and apply all data
}
\examples{
as_printable(mtcars)
as_printable(iris)
}
