% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{autoReg_sub}
\alias{autoReg_sub}
\title{Perform univariable and multivariable regression and stepwise backward regression automatically}
\usage{
autoReg_sub(
  fit,
  threshold = 0.2,
  uni = FALSE,
  multi = TRUE,
  final = FALSE,
  imputed = FALSE,
  keepstats = FALSE
)
}
\arguments{
\item{fit}{An object of class lm or glm}

\item{threshold}{numeric}

\item{uni}{logical whether or not perform univariate regression}

\item{multi}{logical whether or not perform multivariate regression}

\item{final}{logical whether or not perform stepwise backward elimination}

\item{imputed}{logical whether or not include imputed model}

\item{keepstats}{logical whether or not keep statistics}
}
\description{
Perform univariable and multivariable regression and stepwise backward regression automatically
}
