% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{autoRegCox}
\alias{autoRegCox}
\title{Perform univariable and multivariable regression and stepwise backward regression automatically}
\usage{
autoRegCox(
  fit,
  data,
  threshold = 0.2,
  uni = FALSE,
  multi = TRUE,
  final = FALSE,
  keepid = FALSE
)
}
\arguments{
\item{fit}{An object of class coxph}

\item{data}{A data.frame}

\item{threshold}{numeric}

\item{uni}{logical whether or not perform univariate regression}

\item{multi}{logical whether or not perform multivariate regression}

\item{final}{logical whether or not perform stepwise backward elimination}

\item{keepid}{logical whether or not save id column}
}
\description{
Perform univariable and multivariable regression and stepwise backward regression automatically
}
\examples{
require(survival)
require(magrittr)
data(cancer)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
autoRegCox(fit,data=colon) \%>\% myft()
autoRegCox(fit,data=colon,uni=TRUE,threshold=1) \%>\% myft()
autoRegCox(fit,data=colon,uni=TRUE,final=TRUE) \%>\% myft()
data(colon_s,package="finalfit")
fit=coxph(Surv(time,status)~age.factor+sex.factor+obstruct.factor+perfor.factor,data=colon_s)
autoRegCox(fit,data=colon_s,uni=TRUE,threshold=1) \%>\% myft()
autoRegCox(fit,data=colon_s,uni=TRUE,threshold=0.2) \%>\% myft()
}
