% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{baggingElasticNet}
\alias{baggingElasticNet}
\title{This function performs bootstrap bagging of elastic net for R times.}
\usage{
baggingElasticNet(X, y, R = 50)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{R}{The number of bootstrap replicates to perform with default of 50}
}
\value{
This function returns a list of:
1. coef: Coefficients of final bagged model
2. sd.coef: Standard deviation of each coefficients
3. selection.rate: Selection frequency for each bag
}
\description{
This function performs bootstrap bagging of elastic net for R times.
}
