% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{baggingLinear}
\alias{baggingLinear}
\title{This function performs bootstrap bagging R times. R must be given by users.}
\usage{
baggingLinear(X, y, R)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{R}{The number of bootstrap replicates to perform}
}
\value{
This function returns final.model, which is final bagged model.
}
\description{
This function performs bootstrap bagging R times. R must be given by users.
}
