% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{baggingLogistic}
\alias{baggingLogistic}
\title{This function performs bootstrap bagging R times. R must be given by users.}
\usage{
baggingLogistic(X, y, R, p)
}
\arguments{
\item{X}{A matrix of predictor variable}

\item{y}{A vector of response variable}

\item{R}{The number of bootstrap replicates to perform}

\item{p}{The proportion of train data}
}
\value{
This function returns final.model, which is final bagged model.
}
\description{
This function performs bootstrap bagging R times. R must be given by users.
}
