% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{baggingPrediction}
\alias{baggingPrediction}
\title{This function predict outcomes of test data of bagged logistic model.}
\usage{
baggingPrediction(X, y, R = 50, p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variable}

\item{y}{A vector of response variable}

\item{R}{The number of bootstrap replicates to perform with default of 50}

\item{p}{A proportion of train data with default of 0.8}
}
\value{
If y is binary, it returns a list of:
1. confusion.matrix: A confusion matrix between predicted data and test data
2. accuracy: An accuracy of prediction
If y is continuous, it returns
RMSE: RMSE(Root Mean Square Error) between predicted data and test data.
}
\description{
This function predict outcomes of test data of bagged logistic model.
}
