% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2model.R
\name{beNumeric}
\alias{beNumeric}
\title{Whether a string vector can be converted to numeric}
\usage{
beNumeric(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
A logical vector
}
\description{
Whether a string vector can be converted to numeric
}
