% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCompare.R
\name{bootCompare}
\alias{bootCompare}
\title{Prepare data for analysis with dabestr and bootstrap-coupled estimation of effect size}
\usage{
bootCompare(.data, ..., idx, paired, id.column = NULL)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{...}{Column names in .data}

\item{idx}{A vector containing factors or strings in the grouping columns.}

\item{paired}{logical  If TRUE, the groups are treated as paired samples.}

\item{id.column}{Default NULL. A column name indicating the identity of the datapoint if the data is paired. This must be supplied if paired is TRUE.}
}
\value{
A dabest_effsize object with 10 elements. Details are described in \code{\link[dabestr]{mean_diff}}.
}
\description{
Prepare data for analysis with dabestr and bootstrap-coupled estimation of effect size
This function is a wrapper function for \code{\link[dabestr]{dabest}} and \code{\link[dabestr]{mean_diff}}.
You can select variables for \code{\link[dabestr]{dabest}} function using dplyr::select.
This function also compatible with wide-form data.
}
\examples{
require(dplyr)
require(dabestr)
bootCompare(iris,Species,Sepal.Length) \%>\% plot()
\donttest{
data(acs,package="moonBook")
acs \%>\% select(sex,age) \%>\% bootCompare() \%>\% plot()
data(Anorexia,package="PairedData")
bootCompare(Anorexia,Prior,Post)
bootCompare(Anorexia) \%>\% plot()
data(Blink2,package="PairedData")
Blink2 \%>\% dplyr::select(2:4) \%>\% bootCompare() \%>\% plot()
bootCompare(Blink2,Resting,Straight,Oscillating,paired=FALSE)
}
}
