% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check data
This function checks if input X and y are valid type of data to perform further machine learning process}
\usage{
check_data(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
This function returns a list consisting:
response_type: type of response variable y(binary or continuous)
y: finalized response value of y
}
\description{
Check data
This function checks if input X and y are valid type of data to perform further machine learning process
}
