% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{compress}
\alias{compress}
\title{Compress an object of class gaze}
\usage{
compress(x, xname = "name", ref = 1)
}
\arguments{
\item{x}{an object of class gaze}

\item{xname}{A variable name}

\item{ref}{Numeric Th number to be used as reference}
}
\description{
Compress an object of class gaze
}
\examples{
data(acs,package="moonBook")
x=gaze(sex~.,data=acs,keepid=TRUE)
compress(x)
}
