% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countGroups.R
\name{countGroups}
\alias{countGroups}
\title{Count groups}
\usage{
countGroups(data, yvars)
}
\arguments{
\item{data}{A data.frame}

\item{yvars}{variable names}
}
\value{
An object of class "tibble"
}
\description{
Count groups
}
\examples{
library(moonBook)
countGroups(acs,"sex")
countGroups(acs,c("sex","Dx"))
}
