% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr_formula.R
\name{crr_formula}
\alias{crr_formula}
\title{Competing risk Regression with formula}
\usage{
crr_formula(x, data)
}
\arguments{
\item{x}{fomular time+status~explanatory variables}

\item{data}{data a data.frame}
}
\description{
Competing risk Regression with formula
}
\examples{
data(melanoma,package="boot")
melanoma$status_crr=ifelse(melanoma$status==1,1,ifelse(melanoma$status==2,0,2))
crr_formula(time+status_crr~age+sex+thickness+ulcer,data=melanoma)
}
