% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrFormula.R
\name{crrFormula}
\alias{crrFormula}
\title{Competing Risk Regression with Formula}
\usage{
crrFormula(x, data, ...)
}
\arguments{
\item{x}{formula time+status~explanatory variables}

\item{data}{data a data.frame}

\item{...}{Further arguments to be passed to \code{\link[cmprsk]{crr}}}
}
\value{
An object of class "crr" which is described in \code{\link[cmprsk]{crr}}
}
\description{
Competing Risk Regression with Formula
}
\examples{
data(melanoma,package="boot")
melanoma$status_crr=ifelse(melanoma$status==1,1,ifelse(melanoma$status==2,0,2))
crrFormula(time+status_crr~age+sex+thickness+ulcer,data=melanoma)
}
