% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2flextable.R
\name{df2flextable}
\alias{df2flextable}
\title{Convert data.frame to flextable}
\usage{
df2flextable(
  df,
  vanilla = FALSE,
  fontname = NULL,
  fontsize = 12,
  add.rownames = FALSE,
  even_header = "transparent",
  odd_header = "#5B7778",
  even_body = "#EFEFEF",
  odd_body = "transparent",
  vlines = TRUE,
  colorheader = FALSE,
  digits = 2,
  digitp = 3,
  align_header = "center",
  align_body = "right",
  align_rownames = "left",
  NA2space = TRUE,
  pcol = NULL,
  ...
)
}
\arguments{
\item{df}{A data.frame}

\item{vanilla}{A Logical}

\item{fontname}{Font name}

\item{fontsize}{font size}

\item{add.rownames}{logical. Whether or not include rownames}

\item{even_header}{background color of even_header}

\item{odd_header}{background color of even_header}

\item{even_body}{background color of even_body}

\item{odd_body}{background color of even_body}

\item{vlines}{Logical. Whether or not draw vertical lines}

\item{colorheader}{Logical. Whether or not use color in header}

\item{digits}{integer indicating the number of decimal places}

\item{digitp}{integer indicating the number of decimal places of p values}

\item{align_header}{alignment of header. Expected value is one of 'left', 'right', 'center', 'justify'.}

\item{align_body}{alignment of body. Expected value is one of 'left', 'right', 'center', 'justify'.}

\item{align_rownames}{alignment of rownames. Expected value is one of 'left', 'right', 'center', 'justify'.}

\item{NA2space}{A logical. If true, convert NA value to space}

\item{pcol}{An integer indicating p value. If specified, convert value less than 0.01 to "< 0.001" in given column.}

\item{...}{further arguments to be passed to \code{\link[flextable]{flextable}}}
}
\value{
An object of class "flextable" which is described in \code{\link[flextable]{flextable}}
}
\description{
Convert data.frame to flextable
}
