% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensembleLearning}
\alias{ensembleLearning}
\title{This function is to evaluate performance of ensemble learning models including standard regression(linear, logistic) and regularized regression(ridge, LASSO, elastic net) processes.
On every iteration, this function provides the selection of models to perform. Users can stop fitting models by input 6.}
\usage{
ensembleLearning(X, y, p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{p}{A proportion of train data. Default is 0.8}
}
\value{
If y is binary, it returns a list of
1. model_summary: Count of model used in ensemble learning process
2. confusion.matrix: confusion matrix between final predictions and actual value of test data.\
3. accuracy: overall accuracy based on confusion matrix
If y is continuous, it returns a list of
1. model_summary: Count of model used in ensemble learning process
2. RMSE: RMSE(Root Mean Square Error) between predicted data and test data
3. r_squared: R-squared, also known as coefficient of determination
}
\description{
This function is to evaluate performance of ensemble learning models including standard regression(linear, logistic) and regularized regression(ridge, LASSO, elastic net) processes.
On every iteration, this function provides the selection of models to perform. Users can stop fitting models by input 6.
}
