% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{evaluateBagging}
\alias{evaluateBagging}
\title{This function is to evaluate performance of model obtained by bootstrap bagging process.}
\usage{
evaluateBagging(X, y, model = "standard", R = 50, p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variable}

\item{y}{A vector of response variable}

\item{model}{Regression model specification ("standard, "ridge", "lasso", "elastic.net") with default of "standard"}

\item{R}{The number of bootstrap replicates to perform with default of 50}

\item{p}{A proportion of train data with default of 0.8}
}
\value{
If y is binary, it returns a list of:
1. confusion.matrix: A confusion matrix between predicted data and test data
2. accuracy: An accuracy of prediction
If y is continuous, it returns a list of
1. r_squared: R-squared(Coefficient of Determination)
2. RMSE: RMSE(Root Mean Square Error) between predicted data and test data
}
\description{
This function is to evaluate performance of model obtained by bootstrap bagging process.
}
