% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{evaluateModel}
\alias{evaluateModel}
\title{This function is to evaluate performance of model obtained by single regression process.}
\usage{
evaluateModel(X, y, model = "standard", p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{model}{Regression model specification ("standard, "ridge", "lasso", "elastic.net") with default of "standard"}

\item{p}{The proportion of train data}
}
\value{
If y is binary, it returns a list of
1. confusion.matrix: A confusion matrix between predicted data and test data
2. accuracy: An accuracy of prediction
If y is continuous, it returns a list of
1. r_squared: R-squared(Coefficient of determination)
2. RMSE: RMSE(Root Mean Square Error) between predicted data and test data.
}
\description{
This function is to evaluate performance of model obtained by single regression process.
}
