% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPlot2.R
\name{expectedPlot2}
\alias{expectedPlot2}
\title{Draw an expected plot for a numeric predictor}
\usage{
expectedPlot2(fit, xname = NULL, maxy.lev = 5, median = TRUE, se = FALSE, ...)
}
\arguments{
\item{fit}{An object of class "coxph"}

\item{xname}{Character Name of explanatory variable to plot}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}

\item{median}{Logical}

\item{se}{logical Whether or not show se}

\item{...}{further arguments to be passed to plot.survfit}
}
\value{
No return value, called for side effects
}
\description{
Select cutpoint for a numeric predictor with maxstat.test() and draw survival plot with this cutpoint
}
\examples{
library(survival)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~age+sex,data =colon)
expectedPlot2(fit,xname="age")
}
