% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{extractHR2}
\alias{extractHR2}
\title{Extract Harzard Ratio}
\usage{
extractHR2(x, digits = 4)
}
\arguments{
\item{x}{An object of class coxph}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Extract Harzard Ratio
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
extractHR2(fit)
}
