% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{extractOR2}
\alias{extractOR2}
\title{Extract Odds Ratio}
\usage{
extractOR2(x, method = "likelihood", digits = 4)
}
\arguments{
\item{x}{An object of class glm}

\item{method}{character choices are one of the c("likelihood","wald")}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Extract Odds Ratio
}
\examples{
data(cancer,package="survival")
x=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
extractOR2(x)
}
