% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fill vector with lead value}
\usage{
fill(x, what = c("", NA))
}
\arguments{
\item{x}{a vector}

\item{what}{Values to be filled}
}
\description{
Fill vector with lead value
}
\examples{
x=rep(1:5,each=3)
x=removeDup(x,NA)
fill(x)
}
