% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{filldown}
\alias{filldown}
\title{filldown vector with lead value}
\usage{
filldown(x, what = c("", NA))
}
\arguments{
\item{x}{a vector}

\item{what}{Values to be filled}
}
\value{
A vector with the same class as x
}
\description{
filldown vector with lead value
}
\examples{
x=rep(1:5,each=3)
x=removeDup(x,NA)
filldown(x)
}
