% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{find1stDup}
\alias{find1stDup}
\title{Find first duplicated position}
\usage{
find1stDup(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A logical vector
}
\description{
Find first duplicated position
}
\examples{
x=rep(1:5,each=3)
which(find1stDup(x))
}
