% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{findDup}
\alias{findDup}
\title{Find duplicated term}
\usage{
findDup(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector
}
\description{
Find duplicated term
}
\examples{
x=rep(1:5,each=3)
findDup(x)
which(!findDup(x))
}
