% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{fit2final}
\alias{fit2final}
\title{Make final model using stepwise backward elimination}
\usage{
fit2final(fit, threshold = 0.2)
}
\arguments{
\item{fit}{An object of class coxph}

\item{threshold}{Numeric}
}
\description{
Make final model using stepwise backward elimination
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
final=fit2final(fit)
fit2summary(final)
}
