% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze.coxph.R
\name{fit2lik}
\alias{fit2lik}
\title{extract likelihood information with a coxph object}
\usage{
fit2lik(x)
}
\arguments{
\item{x}{An object of class "coxph" or "survreg"}
}
\value{
A string
}
\description{
extract likelihood information with a coxph object
}
\examples{
library(survival)
fit=coxph(Surv(time,status) ~rx,data=anderson)
fit2lik(fit)
}
