% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2model.R
\name{fit2model}
\alias{fit2model}
\title{Restore fit model data containing AsIs expressions}
\usage{
fit2model(fit)
}
\arguments{
\item{fit}{An object of class "lm", "glm" or "coxph"}
}
\value{
An object of class "data.frame"
}
\description{
Restore fit model data containing AsIs expressions
}
\examples{
require(survival)
pbc$status2=ifelse(pbc$status==2,1,0)
fit=coxph(Surv(time,status2)~age+log(bili),data=pbc)
fit2model(fit)
}
