% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{fit2multi}
\alias{fit2multi}
\title{Make multivariate regression model by selecting univariate models with p.value below threshold}
\usage{
fit2multi(fit, threshold = 0.2)
}
\arguments{
\item{fit}{An object of class coxph}

\item{threshold}{Numeric}
}
\description{
Make multivariate regression model by selecting univariate models with p.value below threshold
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
fit2multi(fit)
}
