% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2newdata.R
\name{fit2newdata}
\alias{fit2newdata}
\title{Make a new data of mean value or most frequent value}
\usage{
fit2newdata(fit, xnames = NULL, maxy.lev = 5, digits = 1)
}
\arguments{
\item{fit}{An object of class "coxph"}

\item{xnames}{character Names of explanatory variable to plot}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}

\item{digits}{integer indicating the number of decimal places}
}
\value{
A data.frame
}
\description{
Make a new data of mean value or most frequent value
}
\examples{
require(survival)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~rx+sex+age,data=colon)
fit=coxph(Surv(time,status)~rx+age+strata(sex),data=colon)
fit2newdata(fit)
}
