% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitElasticNet}
\alias{fitElasticNet}
\title{This function fits elastic net model.}
\usage{
fitElasticNet(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted elastic net model
}
\description{
This function fits elastic net model.
}
