% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitElasticNetModel}
\alias{fitElasticNetModel}
\title{This function fit elastic net model.}
\usage{
fitElasticNetModel(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted elastic net model
}
\description{
This function fit elastic net model.
}
