% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitLasso}
\alias{fitLasso}
\title{This function fits LASSO regression model.}
\usage{
fitLasso(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted LASSO regression model
}
\description{
This function fits LASSO regression model.
}
