% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitLassoModel}
\alias{fitLassoModel}
\title{This function fit lasso model.}
\usage{
fitLassoModel(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted lasso model
}
\description{
This function fit lasso model.
}
