% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitLinearModel}
\alias{fitLinearModel}
\title{This function fit linear model.}
\usage{
fitLinearModel(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
fitted linear model
}
\description{
This function fit linear model.
}
