% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitLogisticModel}
\alias{fitLogisticModel}
\title{This function fit logistic model.}
\usage{
fitLogisticModel(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted logistic model
}
\description{
This function fit logistic model.
}
