% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitRandomForestModel}
\alias{fitRandomForestModel}
\title{This function fit random forest model}
\usage{
fitRandomForestModel(X, y, ntree = 500)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{ntree}{The number of trees to grow}
}
\value{
Fitted random forest model
}
\description{
This function fit random forest model
}
