% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitRidge}
\alias{fitRidge}
\title{This function fits ridge regression model.}
\usage{
fitRidge(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted ridge regression model
}
\description{
This function fits ridge regression model.
}
