% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitRidgeModel}
\alias{fitRidgeModel}
\title{This function fit ridge model.}
\usage{
fitRidgeModel(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\value{
Fitted ridge model
}
\description{
This function fit ridge model.
}
