% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitStandard}
\alias{fitStandard}
\title{This function fits standard regression model.}
\usage{
fitStandard(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}
}
\description{
This function fits standard regression model.
}
