% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteraction.R
\name{getInteraction}
\alias{getInteraction}
\title{Get interaction data from data}
\usage{
getInteraction(name, data)
}
\arguments{
\item{name}{a string with interaction term}

\item{data}{a data.frame}
}
\value{
An object of class "data.frame"
}
\description{
Get interaction data from data
}
\examples{
data(acs,package="moonBook")
getInteraction("TC:Dx:sex",data=acs)
}
