% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcmprsk2.R
\name{ggcmprisk2}
\alias{ggcmprisk2}
\title{Compare cumulative incidence to th Kaplan-Meier estimate}
\usage{
ggcmprisk2(
  timevar,
  statusvar,
  data,
  id = c("disease", "other"),
  xpos = c(2, 2),
  ypos = c(0.25, 0.7),
  ylabs = NULL,
  xlab = NULL,
  label = NULL,
  plot = TRUE
)
}
\arguments{
\item{timevar}{a name of time variable}

\item{statusvar}{a name of status variable. coded as one of c(0,1,2)}

\item{data}{A data.frame}

\item{id}{Character vector of length2}

\item{xpos}{numeric x-axis position of label}

\item{ypos}{numeric y-axis position of label}

\item{ylabs}{string vector of length 2. y axis labels}

\item{xlab}{A character. The x-axis label}

\item{label}{string vector of length 2. Label names}

\item{plot}{logical Whether or not print plot}
}
\value{
A list
\describe{
  \item{df}{A long data.frame consist of time, est, upper,lower, id, method}
  \item{df3}{A data.frame for label consist of x, y, label, id}
  \item{p}{A ggplot object}
}
}
\description{
Compare cumulative incidence to th Kaplan-Meier estimate
}
\examples{
require(dplyr)
data(prostateSurvival,package="asaur")
prostateHighRisk <- prostateSurvival \%>\%
  filter(grade=="poor" & stage=="T2",ageGroup=="80+")
ggcmprisk2("survTime/12","status",data=prostateHighRisk,
     id=c("prostate cancer","other causes"))
}
