% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcmprsk.R
\name{ggcmprsk}
\alias{ggcmprsk}
\title{Draw Cumulative Incidence Curves for Competing Risks}
\usage{
ggcmprsk(fit, cnames = NULL, se = FALSE)
}
\arguments{
\item{fit}{An  object of class cuminc}

\item{cnames}{character vector}

\item{se}{logical whether or not show confidence interval}
}
\description{
Draw Cumulative Incidence Curves for Competing Risks
}
\examples{
library(cmprsk)
data(melanoma,package="boot")
melanoma$status1 = ifelse(melanoma$status==1,1,ifelse(melanoma$status==2,0,2))
fit=cuminc(melanoma$time/365,melanoma$status1)
ggcmprsk(fit,c("melanoma","others"))
}
