% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2newdata.R
\name{is.mynumeric}
\alias{is.mynumeric}
\title{Decide whether a vector can be treated as a numeric variable}
\usage{
is.mynumeric(x, maxy.lev = 5)
}
\arguments{
\item{x}{A vector}

\item{maxy.lev}{An integer indicating the maximum number of unique values of a numeric variable be treated as a categorical variable}
}
\value{
A logical value
}
\description{
Decide whether a vector can be treated as a numeric variable
}
\examples{
x=1:5
is.mynumeric(x)
x=1:13
is.mynumeric(x)
}
