% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPlot.R
\name{label_parse}
\alias{label_parse}
\title{takes the breaks as input and returns labels as output}
\usage{
label_parse(breaks)
}
\arguments{
\item{breaks}{character}
}
\value{
a character vector
}
\description{
takes the breaks as input and returns labels as output
}
