% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogplot.R
\name{loglogplot}
\alias{loglogplot}
\title{Draw log-log plot}
\usage{
loglogplot(fit, xname = NULL, main = NULL, labels = NULL, maxy.lev = 5)
}
\arguments{
\item{fit}{An object of class "coxph"}

\item{xname}{A character Name of explanatory variable to plot}

\item{main}{String Title of plot}

\item{labels}{String vector Used as legend in legend}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}
}
\value{
No return value, called for side effects
}
\description{
Draw log-log plot
}
\examples{
require(survival)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~x,data=leukemia)
loglogplot(fit)
}
