% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gaze.R
\name{maxnchar}
\alias{maxnchar}
\title{Retur maximum character number except NA}
\usage{
maxnchar(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Retur maximum character number except NA
}
\examples{
x=c(1,2,"sadf",NA)
maxnchar(x)
data(acs,package="moonBook")
lapply(acs,maxnchar)
}
