% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{modelsSummary}
\alias{modelsSummary}
\title{Makes table summarizing list of models}
\usage{
modelsSummary(fitlist, show.lik = FALSE)
}
\arguments{
\item{fitlist}{A list of objects of class "coxph"}

\item{show.lik}{logical Whether or not show likelihood test results}
}
\value{
No return value, called for side effects
}
\description{
Makes table summarizing list of models
}
\examples{
library(survival)
fit1=coxph(Surv(time,status) ~rx,data=anderson)
fit2=coxph(Surv(time,status) ~rx+logWBC,data=anderson)
fit3=coxph(Surv(time,status) ~rx*logWBC,data=anderson)
fitlist=list(fit1,fit2,fit3)
modelsSummary(fitlist)
}
