% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze_sub.R
\name{my.chisq.test2}
\alias{my.chisq.test2}
\title{Statistical test for categorical variables
Statistical test for categorical variables}
\usage{
my.chisq.test2(x, y, catMethod = 2, all = FALSE)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{catMethod}{An integer indicating methods for categorical variables.
Possible values in methods are
\describe{
   \item{0}{Perform chisq.test first. If warning present, perform fisher test}
   \item{1}{Perform chisq.test without continuity correction}
   \item{2}{Perform chisq.test with continuity correction}
   \item{3}{perform fisher.test}
   \item{4}{perform prop.trend test}
}
Default value is 2.}

\item{all}{A logical}
}
\description{
Statistical test for categorical variables
Statistical test for categorical variables
}
\examples{
library(moonBook)
x=acs$sex
y=acs$Dx
my.chisq.test2(x,y)
}
