% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{mycphSimple}
\alias{mycphSimple}
\title{Fit Simple Proportional Hazards Regression Model}
\usage{
mycphSimple(fit, threshold = 0.2, digits = 2)
}
\arguments{
\item{fit}{An object of class coxph}

\item{threshold}{numeric p-value threshold to enter multiple model}

\item{digits}{interger indicating the position decimal place}
}
\description{
Fit Simple Proportional Hazards Regression Model
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
}
