% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_printable.R
\name{myformat}
\alias{myformat}
\title{Convert data.frame to printable format}
\usage{
myformat(x, showid = FALSE, digits = 3)
}
\arguments{
\item{x}{A data.frame}

\item{showid}{logical if TRUE, show id}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
A data.frame
}
\description{
Convert data.frame to printable format
}
\examples{
fit=lm(mpg~wt*hp,data=mtcars)
gaze(fit) \%>\% myformat()
}
