% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze.R
\name{myft}
\alias{myft}
\title{Convert data.frame into flextable}
\usage{
myft(x, vanilla = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{vanilla}{logical}
}
\description{
Convert data.frame into flextable
}
\examples{
data(acs,package="moonBook")
library(magrittr)
gaze(acs) \%>\% myft()
gaze(sex~.,acs) \%>\% myft()
gaze(sex+Dx~.,data=acs,show.p=TRUE,show.total=TRUE,show.n=TRUE,shiw.missing=TRUE) \%>\% myft()
gaze(Dx+sex~cardiogenicShock,data=acs,show.p=TRUE) \%>\% myft()
gaze(Dx+sex+HBP~cardiogenicShock,data=acs,show.p=TRUE) \%>\% myft()
}
