% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySummary.R
\name{mySummary}
\alias{mySummary}
\alias{mySummary.formula}
\alias{mySummary.data.frame}
\title{Produce table for descriptive statistics}
\usage{
mySummary(x, ...)

\method{mySummary}{formula}(x, ...)

\method{mySummary}{data.frame}(x, ...)
}
\arguments{
\item{x}{An R object, formula or data.frame}

\item{...}{arguments to be passed to mySummary.data.frame or mySummary.formula}
}
\description{
Produce table for descriptive statistics by groups for several variables easily.
Depending on  the nature of these variables, different descriptive statistical
methods were used(t-test, ANOVA,Kruskal-Wallis, chisq, Fisher,...)
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: S3 method for formula

\item \code{data.frame}: default S3 method
}}

\examples{
library(moonBook)
library(magrittr)
mySummary(acs)
mySummary(~age+sex,data=acs)
mySummary(sex~.,data=acs,digits=1,method=1,show.p=TRUE) \%>\% myft()
\donttest{
mySummary(sex~age+Dx,data=acs)
mySummary(EF~.,data=acs) \%>\% myft()
mySummary(sex+Dx~.,data=acs,show.p=TRUE) \%>\% myft()
mySummary(sex+Dx~.,data=acs)
mySummary(Dx+sex~cardiogenicShock,data=acs,show.p=TRUE) \%>\% myft()
mySummary(Dx+sex+HBP~cardiogenicShock,data=acs,show.p=TRUE)
mySummary(~mpg+cyl,data=mtcars)
mySummary(~.,data=mtcars)
mySummary(cyl~.,data=mtcars,show.p=TRUE)
mySummary(hp~.,data=mtcars)
mySummary(cyl+am~.,data=mtcars)
}
}
