% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{mySummary_sub}
\alias{mySummary_sub}
\title{Summary function for categorical/continuous variable}
\usage{
mySummary_sub(data, x, y = NULL, max.ylev = 5, ...)
}
\arguments{
\item{data}{A data.frame}

\item{x}{A name of categorical/continuous vector}

\item{y}{A name of vector, either continuous or categorical}

\item{max.ylev}{max.ylev An integer indicating the maximum number of levels of grouping variable ('y').
If a column have unique values less than max.ylev it is treated as a categorical variable. Default value is 5.}

\item{...}{Further arguments to be passed to mySummaryCont() or mySummaryCat()}
}
\description{
Summary function for categorical/continuous variable
}
\examples{
require(moonBook)
mySummary_sub(acs,"age")
mySummary_sub(acs,"sex")
mySummary_sub(acs,"age","EF")
mySummary_sub(acs,"sex","EF")
mySummary_sub(acs,"age","Dx")
mySummary_sub(acs,"sex","Dx")
mySummary_sub(iris,"Species","Sepal.Length")
}
