% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysurvregSimple.R
\name{mysurvregSimple}
\alias{mysurvregSimple}
\title{Fit Simple AFT Model}
\usage{
mysurvregSimple(fit, threshold = 0.2, digits = 2, mode = 1)
}
\arguments{
\item{fit}{An object of class survreg}

\item{threshold}{numeric p-value threshold to enter multiple model}

\item{digits}{integer indicating the position decimal place}

\item{mode}{integer}
}
\value{
An object of class "data.frame"
}
\description{
Fit Simple AFT Model
}
\examples{
require(survival)
data(cancer)
fit=survreg(Surv(time,status)~rx+age+strata(sex)+obstruct+perfor,data=colon)
mysurvregSimple(fit)
}
