% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedPlot.R
\name{num2factor}
\alias{num2factor}
\title{Convert a numeric column in a data.frame to a factor}
\usage{
num2factor(data, call, name, no = 3)
}
\arguments{
\item{data}{A data.frame}

\item{call}{a function call}

\item{name}{character Name of numeric column}

\item{no}{numeric}
}
\value{
A data.frame
}
\description{
Convert a numeric column in a data.frame to a factor
}
\examples{
num2factor(anderson,name="logWBC")
library(survival)
fit=coxph(Surv(time,status)~logWBC+rx,data=anderson)
num2factor(anderson,call=fit$call,name="logWBC",no=2)
}
