% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze_sub.R
\name{num2stat}
\alias{num2stat}
\title{Summarize numeric vector to statistical summary}
\usage{
num2stat(x, digits = 1, method = 1, p = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{integer indicating the number of decimal places}

\item{method}{An integer indicating methods for continuous variables. Possible values in methods are
1 forces analysis as normal-distributed
2 forces analysis as continuous non-normal
3 performs a Shapiro-Wilk test or nortest::ad.test to decide between normal or non-normal
Default value is 1.}

\item{p}{A numeric}
}
\value{
A character vector of length 1
}
\description{
Summarize numeric vector to statistical summary
}
\examples{
library(moonBook)
num2stat(acs$age)
num2stat(acs$age,method=2)
}
