% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEplot.R
\name{OEplot}
\alias{OEplot}
\title{Draw an Observed vs Expected plot}
\usage{
OEplot(fit, xname = NULL, maxy.lev = 5)
}
\arguments{
\item{fit}{An object of class "coxph"}

\item{xname}{A character Name of explanatory variable to plot}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}
}
\value{
No return value, called for side effects
}
\description{
Draw an Observed vs Expected plot
}
\examples{
library(survival)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~rx,data=colon)
OEplot(fit)
}
