% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{p2character2}
\alias{p2character2}
\title{Change p value to string}
\usage{
p2character2(x, digits = 3, add.p = TRUE)
}
\arguments{
\item{x}{a numeric}

\item{digits}{interger indicating decimal place}

\item{add.p}{logical}
}
\description{
Change p value to string
}
