% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{prescreening}
\alias{prescreening}
\title{Select K most informative predictors}
\usage{
prescreening(X, y, K = 1)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{K}{The number of informative variables}
}
\value{
The names of K most informative variables
}
\description{
Select K most informative predictors
}
