% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{print.autoReg}
\alias{print.autoReg}
\title{S3 method print for an object of class autoReg}
\usage{
\method{print}{autoReg}(x, ...)
}
\arguments{
\item{x}{An object of class autoReg}

\item{...}{Further arguments}
}
\description{
S3 method print for an object of class autoReg
}
\examples{
data(cancer,package="survival")
fit=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
autoReg(fit)
}
