% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gaze.R
\name{print.gaze}
\alias{print.gaze}
\title{S3 method print for an object of class gaze}
\usage{
\method{print}{gaze}(x, ...)
}
\arguments{
\item{x}{An object of class gaze}

\item{...}{Further arguments}
}
\value{
No return value, called for side effects
}
\description{
S3 method print for an object of class gaze
}
\examples{
data(acs,package="moonBook")
x=gaze(acs,show.n=TRUE,show.missing=TRUE)
gaze(sex~.,acs,show.p=TRUE,show.n=TRUE,show.missing=TRUE,show.total=TRUE)
\donttest{
gaze(Dx+sex~.,acs,show.p=TRUE)
gaze(sex+Dx+HBP~.,acs,show.p=TRUE)
}
}
