% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{print.modelPlot}
\alias{print.modelPlot}
\title{S3 method for an class modelPlot}
\usage{
\method{print}{modelPlot}(x, ...)
}
\arguments{
\item{x}{An object of class modelPlot}

\item{...}{Further arguments to be passed to plot()}
}
\description{
S3 method for an class modelPlot
}
