% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printdf.R
\name{printdf}
\alias{printdf}
\title{Print function for data.frame}
\usage{
printdf(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
No return value, called for side effects
}
\description{
Print function for data.frame
}
\examples{
x=mtcars[1:5,1:5]
printdf(x)
}
