% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualPlot.R
\name{residualPlot}
\alias{residualPlot}
\title{Draw a residual plot with an object of class coxph}
\usage{
residualPlot(fit, type = "martingale")
}
\arguments{
\item{fit}{An object of class coxph}

\item{type}{character One of the c("martingale","deviance","score","schoenfeld",
"dfbeta","dfbetas","scaledsch","partial"). Default value is "martingale".}
}
\value{
A patchwork object
}
\description{
Draw a residual plot with an object of class coxph
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status==2)~log(bili)+age+cluster(edema),data=pbc)
residualPlot(fit)
fit=coxph(Surv(time,status==2)~age,data=pbc)
residualPlot(fit)
residualPlot(fit,"partial")
}
