% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2model.R
\name{restoreData2}
\alias{restoreData2}
\title{restore data with I() in column name}
\usage{
restoreData2(df)
}
\arguments{
\item{df}{An object of class "data.frame"}
}
\value{
An object of class "data.frame"
}
\description{
restore data with I() in column name
}
