% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2model.R
\name{restoreData3}
\alias{restoreData3}
\title{restore data with operator in column name}
\usage{
restoreData3(df, changeLabel = FALSE)
}
\arguments{
\item{df}{An object of class "data.frame"}

\item{changeLabel}{logical}
}
\value{
An object of class "data.frame"
}
\description{
restore data with operator in column name
}
