% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2flextable.R
\name{roundDf}
\alias{roundDf}
\title{Convert numeric columns of data.frame to character}
\usage{
roundDf(df, digits = 2)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Convert numeric columns of data.frame to character
}
