% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabel.R
\name{setLabel}
\alias{setLabel}
\title{Add label to a vector}
\usage{
setLabel(x, label = "")
}
\arguments{
\item{x}{a vector}

\item{label}{string}
}
\value{
a labelled vector
}
\description{
Add label to a vector
}
