% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{shorten}
\alias{shorten}
\title{Shorten an object of class gaze}
\usage{
shorten(x, xname = NULL, ref = 1)
}
\arguments{
\item{x}{an object of class gaze}

\item{xname}{A variable name}

\item{ref}{Numeric Th number to be used as reference}
}
\description{
Shorten an object of class gaze
}
\examples{
data(acs,package="moonBook")
x=gaze(sex~.,data=acs)
shorten(x)
}
