% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata2df.R
\name{strata2df}
\alias{strata2df}
\title{Convert a character vector to a data.frame}
\usage{
strata2df(strata)
}
\arguments{
\item{strata}{A character vector}
}
\value{
A data.frame
}
\description{
Convert a character vector to a data.frame
}
