% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit2df.R
\name{survfit2df}
\alias{survfit2df}
\title{Extract survival data from an object of class "survfit"}
\usage{
survfit2df(fit)
}
\arguments{
\item{fit}{An object of class "survfit"}
}
\value{
A data.frame
}
\description{
Extract survival data from an object of class "survfit"
}
\examples{
library(survival)
data(cancer,package="survival")
fit=survfit(Surv(time,status)~rx+sex+age,data=colon)
survfit2df(fit)
}
