% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2final.R
\name{survreg2final}
\alias{survreg2final}
\title{Make final model using stepwise backward elimination}
\usage{
survreg2final(fit, threshold = 0.2)
}
\arguments{
\item{fit}{An object of class "survreg"}

\item{threshold}{Numeric}
}
\value{
An object of class "survreg" which is described in \code{\link[survival]{survreg}}
}
\description{
Make final model using stepwise backward elimination
}
\examples{
require(survival)
data(cancer)
fit=survreg(Surv(time,status)~rx+age+sex+obstruct+perfor,data=colon)
survreg2final(fit)
}
