% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2multi.R
\name{survreg2multi}
\alias{survreg2multi}
\title{Make multivariable regression model by selecting univariable models with p.value below threshold}
\usage{
survreg2multi(fit, threshold = 0.2)
}
\arguments{
\item{fit}{An object of class "survreg"}

\item{threshold}{Numeric}
}
\value{
An object of class "survreg"
}
\description{
Make multivariable regression model by selecting univariable models with p.value below threshold
}
\examples{
require(survival)
data(cancer)
fit=survreg(Surv(time,status)~rx+age+sex+obstruct+perfor,data=colon)
survreg2multi(fit)
}
