% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{train_test}
\alias{train_test}
\title{Generate test and train data}
\usage{
train_test(X, y, p)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{p}{A proportion of train data}
}
\description{
Generate test and train data
}
