% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{addFitSummary}
\alias{addFitSummary}
\title{Add model summary to an object of class gaze}
\usage{
addFitSummary(df, fit, statsname = "", keepid = FALSE)
}
\arguments{
\item{df}{An object of class gaze}

\item{fit}{An object of class glm or lm}

\item{statsname}{character Name of statistics}

\item{keepid}{logical whether or not save id}
}
\description{
Add model summary to an object of class gaze
}
\examples{
require(survival)
require(magrittr)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~rx+age+sex+nodes+obstruct+perfor,data=colon)
df=autoRegCox(fit,data=colon,uni=FALSE,keepid=TRUE)
final=fit2final(fit)
df \%>\% addFitSummary(final,statsname="HR (final)",keepid=FALSE) \%>\% myft()
}
