% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPlot.R
\name{adjustedPlot}
\alias{adjustedPlot}
\title{Draw an expected plot}
\usage{
adjustedPlot(
  fit,
  xnames = NULL,
  pred.values = list(),
  maxy.lev = 5,
  median = TRUE,
  facet = NULL,
  se = FALSE,
  mark.time = FALSE,
  type = "ggplot",
  ...
)
}
\arguments{
\item{fit}{An object of class "coxph" or "survreg"}

\item{xnames}{Character Names of explanatory variable to plot}

\item{pred.values}{A list A list of predictor values}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}

\item{median}{Logical}

\item{facet}{Character Name of facet variable}

\item{se}{logical Whether or not show se}

\item{mark.time}{logical Whether or not mark time}

\item{type}{Character plot type}

\item{...}{further arguments to be passed to plot.survfit}
}
\value{
A ggplot or no return value(called for side effects)
}
\description{
Draw an expected plot
}
\examples{
library(survival)
fit=coxph(Surv(time,status)~rx+logWBC,data=anderson)
adjustedPlot(fit)
adjustedPlot(fit,xnames="rx",se=TRUE,type="plot")
adjustedPlot(fit,xnames="rx",se=TRUE)
data(cancer,package="survival")
fit=coxph(Surv(time,status)~rx+strata(sex)+age+differ,data =colon)
adjustedPlot(fit,xnames=c("sex"))
adjustedPlot(fit,xnames=c("sex"),pred.values=list(age=58,differ=3))
adjustedPlot(fit,xnames=c("sex","rx"),facet="sex")
adjustedPlot(fit,xnames=c("rx","sex","differ"),facet=c("sex","rx"),se=TRUE)
}
