% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPlot.survreg.R
\name{adjustedPlot.survreg}
\alias{adjustedPlot.survreg}
\title{Draw predicted survival curve with an object survreg}
\usage{
adjustedPlot.survreg(
  x,
  xnames = NULL,
  pred.values = list(),
  maxy.lev = 5,
  median = TRUE,
  newdata = NULL,
  fillalpha = 0.3,
  se = FALSE,
  legend.position = c(0.8, 0.9),
  addKM = FALSE
)
}
\arguments{
\item{x}{An object of class survreg}

\item{xnames}{Character Names of explanatory variable to plot}

\item{pred.values}{A list A list of predictor values}

\item{maxy.lev}{Integer Maximum unique length of a numeric variable to be treated as categorical variables}

\item{median}{Logical}

\item{newdata}{A data.frame or NULL}

\item{fillalpha}{Numeric alpha value for geom_ribbon}

\item{se}{logical Whether or not show se}

\item{legend.position}{legend position}

\item{addKM}{logical Whether or not add KM}
}
\value{
A ggplot
}
\description{
Draw predicted survival curve with an object survreg
}
\examples{
library(survival)
x=survreg(Surv(time, status) ~ 1, data=anderson,dist="exponential")
adjustedPlot(x,addKM=TRUE)
x=survreg(Surv(time, status) ~ 1, data=lung,dist="weibull")
adjustedPlot(x,addKM=TRUE)
x=survreg(Surv(time, status) ~ rx, data=anderson,dist="exponential")
adjustedPlot(x)
x=survreg(Surv(time, status) ~ ph.ecog + age + strata(sex), lung)
adjustedPlot(x)
}
