% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{autoReg}
\alias{autoReg}
\title{Perform univariable and multivariable regression and stepwise backward regression automatically}
\usage{
autoReg(
  fit,
  threshold = 0.2,
  uni = TRUE,
  multi = TRUE,
  final = FALSE,
  imputed = FALSE
)
}
\arguments{
\item{fit}{An object of class lm or glm}

\item{threshold}{numeric}

\item{uni}{logical whether or not perform univariate regression}

\item{multi}{logical whether or not perform multivariate regression}

\item{final}{logical whether or not perform stepwise backward elimination}

\item{imputed}{logical whether or not include imputed model}
}
\description{
Perform univariable and multivariable regression and stepwise backward regression automatically
}
\examples{
library(survival)
data(cancer)
fit=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
autoReg(fit)
autoReg(fit,uni=FALSE,final=TRUE)
autoReg(fit,uni=FALSE,imputed=TRUE)
fit=lm(mpg~wt*hp+am+I(wt^2),data=mtcars)
autoReg(fit,final=TRUE)
autoReg(fit,imputed=TRUE)
}
