% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycphSimple.R
\name{autoRegCox}
\alias{autoRegCox}
\title{perform automatc regression for a class of coxph}
\usage{
autoRegCox(
  x,
  data,
  threshold = 0.2,
  uni = FALSE,
  multi = TRUE,
  final = FALSE,
  imputed = FALSE,
  keepstats = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class coxph}

\item{data}{A data.frame}

\item{threshold}{numeric}

\item{uni}{logical whether or not perform univariate regression}

\item{multi}{logical whether or not perform multivariate regression}

\item{final}{logical whether or not perform stepwise backward elimination}

\item{imputed}{logical whether or not perform multiple imputation}

\item{keepstats}{logical whether or not keep statistic}

\item{...}{Further arguments to be passed to gaze()}
}
\description{
perform automatc regression for a class of coxph
}
\examples{
require(survival)
require(magrittr)
data(cancer)
fit=coxph(Surv(time,status)~rx+age+sex+obstruct+perfor,data=colon)
autoReg(fit,data=colon)
autoReg(fit,data=colon,uni=TRUE,threshold=1)
autoReg(fit,data=colon,uni=TRUE,final=TRUE) \%>\% myft()
data(colon_s,package="finalfit")
fit=coxph(Surv(time,status)~age.factor+sex.factor+obstruct.factor+perfor.factor,data=colon_s)
autoReg(fit,data=colon_s,uni=TRUE,threshold=1)
autoReg(fit,data=colon_s,uni=TRUE,imputed=TRUE)
}
