% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagging.R
\name{baggingElasticNet}
\alias{baggingElasticNet}
\title{This function performs bootstrap bagging of elastic model R times. R must be given by users.}
\usage{
baggingElasticNet(X, y, R, p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{R}{The number of bootstrap replicates to perform}
}
\value{
This function returns a list of:
1. final.coef: Coefficients of final bagged model
2. var.score: Variance importance score
}
\description{
This function performs bootstrap bagging of elastic model R times. R must be given by users.
}
