% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrFormula.R
\name{crrFormula}
\alias{crrFormula}
\title{Competing Risk Regression with Formula}
\usage{
crrFormula(x, data)
}
\arguments{
\item{x}{fomular time+status~explanatory variables}

\item{data}{data a data.frame}
}
\description{
Competing Risk Regression with Formula
}
\examples{
data(melanoma,package="boot")
melanoma$status_crr=ifelse(melanoma$status==1,1,ifelse(melanoma$status==2,0,2))
crrFormula(time+status_crr~age+sex+thickness+ulcer,data=melanoma)
}
