% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaze_sub.R
\name{descNum}
\alias{descNum}
\title{Make description for numeric summary}
\usage{
descNum(method = 1, p = NULL)
}
\arguments{
\item{method}{numeric}

\item{p}{A numeric or NULL}
}
\description{
Make description for numeric summary
}
