% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensemble.R
\name{ensembleLearning}
\alias{ensembleLearning}
\title{This function predicts the outcome of test data in ensemble learning of models including linear, logistic, ridge, lasso, elastic net, and random forest.
On every iteration, this function provides the selection of models to perform. Users can stop fitting models by input 7.}
\usage{
ensembleLearning(X, y, p = 0.8)
}
\arguments{
\item{X}{A matrix of predictor variables}

\item{y}{A vector of response variable}

\item{p}{A proportion of train data. Default is 0.8}
}
\value{
If y is binary, it returns a list of
1. confusion.matrix: confusion matrix between final predictions and actual value of test data.
2. accuracy: overall accuracy based on confusion matrix
If y is continuous, it returns a RMSE value between final predictions and actual value of test data.
}
\description{
This function predicts the outcome of test data in ensemble learning of models including linear, logistic, ridge, lasso, elastic net, and random forest.
On every iteration, this function provides the selection of models to perform. Users can stop fitting models by input 7.
}
