% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{fit2summary}
\alias{fit2summary}
\title{Summarise statistics with a model or model list}
\usage{
fit2summary(fit, ...)
}
\arguments{
\item{fit}{An object of class lm or glm of filtlist}

\item{...}{Further argument to be passed to fit2stats}
}
\description{
Summarise statistics with a model or model list
}
\examples{
library(survival)
data(cancer)
fit=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
fit2summary(fit)
fitlist=fit2list(fit)
fit2summary(fitlist)
}
