% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySummary.R
\name{getGroupNames}
\alias{getGroupNames}
\title{Get sepcific group name}
\usage{
getGroupNames(data, yvars)
}
\arguments{
\item{data}{A data.frame}

\item{yvars}{variable names}
}
\description{
Get sepcific group name
}
\examples{
library(moonBook)
getGroupNames(acs,"sex")
getGroupNames(acs,c("sex","Dx"))
}
