% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{imputedReg}
\alias{imputedReg}
\title{Make a multiple imputed model}
\usage{
imputedReg(fit, data = NULL, m = 20, seed = 1234, digits = 2, ...)
}
\arguments{
\item{fit}{An object of class lm or glm}

\item{data}{a data.frame}

\item{m}{Number of multiple imputations. The default is m=5.}

\item{seed}{An integer that is used as argument by the set.seed() for offsetting the random number generator.}

\item{digits}{Integer indicating the number of decimal place}

\item{...}{Further argument to be passed to mice}
}
\description{
Make a multiple imputed model
}
\examples{
library(survival)
data(cancer,package="survival")
fit=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
imputedReg(fit)
fit=coxph(Surv(time,status)~rx+age+sex+nodes+obstruct+perfor,data=colon)
imputedReg(fit)
}
