% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{modelPlot}
\alias{modelPlot}
\title{Draw coefficients/odds ratio/hazard ratio plot}
\usage{
modelPlot(fit, widths = NULL, change.pointsize = TRUE)
}
\arguments{
\item{fit}{An object of calss glm}

\item{widths}{Numeric vector}

\item{change.pointsize}{logical Whether or not change point size}
}
\description{
Draw coefficients/odds ratio/hazard ratio plot
}
\examples{
fit=lm(mpg~wt*hp+am,data=mtcars)
modelPlot(fit,width=c(1,0,2,3))
fit=lm(Sepal.Width~Species*Sepal.Length,data=iris)
modelPlot(fit)
data(cancer,package="survival")
fit=glm(status~rx+age+sex+nodes+obstruct+perfor,data=colon,family="binomial")
modelPlot(fit)
data(colon_s,package="finalfit")
fit=glm(mort_5yr~age.factor+sex.factor+obstruct.factor+perfor.factor,data=colon_s,family="binomial")
modelPlot(fit)
library(survival)
fit=coxph(Surv(time,status)~age+sex+obstruct+perfor,data=colon)
modelPlot(fit)
fit=coxph(Surv(time,status)~age.factor+sex.factor+obstruct.factor+perfor.factor,data=colon_s)
modelPlot(fit)
}
