% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReg.R
\name{printdf}
\alias{printdf}
\title{Print function for data.frame}
\usage{
printdf(x, showid = FALSE)
}
\arguments{
\item{x}{A data.frame}

\item{showid}{logical if TRUE, show id}
}
\description{
Print function for data.frame
}
