% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualPlot.R
\name{residualPlot}
\alias{residualPlot}
\title{Draw a residual plot with an object of class coxph}
\usage{
residualPlot(
  fit,
  type = "martingale",
  vars = NULL,
  ncol = 2,
  show.point = TRUE,
  se = TRUE
)
}
\arguments{
\item{fit}{An object of class coxph}

\item{type}{character One of the c("martingale","deviance","score","schoenfeld",
"dfbeta","dfbetas","scaledsch","partial"). Default value is "martingale".}

\item{vars}{character Names of variables to plot. default value is NULL}

\item{ncol}{numeric number of columns}

\item{show.point}{logical Whether or not show point}

\item{se}{logical Whether or not show se}
}
\value{
A patchwork object
}
\description{
Draw a residual plot with an object of class coxph
}
\examples{
require(survival)
data(cancer)
fit=coxph(Surv(time,status==2)~log(bili)+age+cluster(edema),data=pbc)
residualPlot(fit)
residualPlot(fit,var="age")
fit=coxph(Surv(time,status==2)~age,data=pbc)
residualPlot(fit)
residualPlot(fit,"partial")
fit=coxph(Surv(time,status)~rx+sex+logWBC,data=anderson)
residualPlot(fit,ncol=3)
}
