% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2model.R
\name{restoreData3}
\alias{restoreData3}
\title{restore data with operator in column name}
\usage{
restoreData3(df, changeLabel = FALSE)
}
\arguments{
\item{df}{a data.frame}

\item{changeLabel}{logical}
}
\description{
restore data with operator in column name
}
